fix(cocod): enforce protected session start policy
