+++ packages/cocod/src/v1/schema.ts
@@ -1,0 +1,345 @@
+import type { CocodStatus } from '../runtime.js';
+
+/** Runtime validator paired with the JSON Schema emitted for the same document. */
+export interface RuntimeSchema<T> {
+  readonly name: string;
+  readonly jsonSchema: Readonly<Record<string, unknown>>;
+  parse(value: unknown): T;
+}
+
+/** Stable error codes that v1 clients may branch on. */
+export const V1_ERROR_CODES = [
+  'unauthenticated',
+  'forbidden',
+  'invalid_request',
+  'not_found',
+  'method_not_allowed',
+  'internal_error',
+  'invalid_idempotency_key',
+  'idempotency_key_conflict',
+  'idempotency_capacity_exceeded',
+  'wallet_already_configured',
+  'wallet_not_configured',
+  'passphrase_required',
+  'wallet_unlock_failed',
+  'session_transition_in_progress',
+  'session_restart_required',
+] as const;
+
+/** Stable machine-readable code carried by every v1 error document. */
+export type V1ErrorCode = (typeof V1_ERROR_CODES)[number];
+
+/** Common v1 error response returned without a legacy envelope. */
+export interface V1ErrorDocument {
+  error: {
+    code: V1ErrorCode;
+    message: string;
+    retryable: boolean;
+    details?: Record<string, unknown>;
+  };
+}
+
+/** Public process-liveness document that contains no Wallet state. */
+export interface HealthDocument {
+  status: 'ok';
+  interfaceVersion: '1';
+}
+
+/** Authenticated Wallet, Seed Access, and Coco Session status document. */
+export interface LifecycleStatusDocument {
+  daemon: {
+    version: string;
+    interfaceVersion: '1';
+  };
+  wallet: {
+    configuredAt: string;
+  } | null;
+  seedAccess: CocodStatus['seedAccess'];
+  cocoSession: CocodStatus['cocoSession'];
+}
+
+/** Network Wallet initialization input; recovery material is always generated by cocod. */
+export interface InitializeWalletRequest {
+  passphrase?: string;
+}
+
+/** Coco Session start input containing optional Wallet-unlocking material. */
+export interface StartSessionRequest {
+  passphrase?: string;
+}
+
+/** Coco Session stop accepts an empty JSON object. */
+export type StopSessionRequest = Record<string, never>;
+
+/** Non-cacheable result returned after cocod durably configures its generated Wallet. */
+export interface InitializeWalletResponseDocument {
+  generatedMnemonic: string;
+  status: LifecycleStatusDocument;
+}
+
+const INTERFACE_VERSION = '1' as const;
+
+interface SchemaNode {
+  readonly jsonSchema: Readonly<Record<string, unknown>>;
+  parse(value: unknown, path: string): unknown;
+}
+
+const rfc3339UtcSchema = stringNode({
+  format: 'date-time',
+  pattern: '^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$',
+});
+const daemonStatusNode = objectNode({
+  version: stringNode(),
+  interfaceVersion: literalNode(INTERFACE_VERSION),
+});
+const walletStatusNode = objectNode({ configuredAt: rfc3339UtcSchema });
+const seedAccessStatusNode = objectNode({
+  state: enumNode(['locked', 'available']),
+  requiresPassphrase: booleanNode(),
+});
+const lastFailureNode = nullableNode(
+  objectNode({
+    code: stringNode(),
+    message: stringNode(),
+    occurredAt: rfc3339UtcSchema,
+  }),
+);
+const cocoSessionStatusNode = objectNode({
+  state: enumNode(['stopped', 'starting', 'running', 'stopping', 'failed']),
+  startedAt: nullableNode(rfc3339UtcSchema),
+  lastFailure: lastFailureNode,
+});
+const lifecycleStatusNode = unionNode([
+  objectNode({
+    daemon: daemonStatusNode,
+    wallet: literalNode(null),
+    seedAccess: literalNode(null),
+    cocoSession: cocoSessionStatusNode,
+  }),
+  objectNode({
+    daemon: daemonStatusNode,
+    wallet: walletStatusNode,
+    seedAccess: seedAccessStatusNode,
+    cocoSession: cocoSessionStatusNode,
+  }),
+]);
+
+/** Runtime schema for a route that accepts no request body. */
+export const noBodySchema = namedSchema<null>('NoBody', literalNode(null));
+
+/** Runtime and generated schema for `GET /health`. */
+export const healthSchema = namedSchema<HealthDocument>(
+  'Health',
+  objectNode({
+    status: literalNode('ok'),
+    interfaceVersion: literalNode(INTERFACE_VERSION),
+  }),
+);
+
+/** Runtime and generated schema for the common v1 error document. */
+export const v1ErrorSchema = namedSchema<V1ErrorDocument>(
+  'Error',
+  objectNode({
+    error: objectNode(
+      {
+        code: enumNode(V1_ERROR_CODES),
+        message: stringNode(),
+        retryable: booleanNode(),
+        details: objectNode({}, { additionalProperties: true }),
+      },
+      { optional: ['details'] },
+    ),
+  }),
+);
+
+/** Runtime and generated schema for authenticated lifecycle status. */
+export const lifecycleStatusSchema = namedSchema<LifecycleStatusDocument>(
+  'LifecycleStatus',
+  lifecycleStatusNode,
+);
+
+/** Runtime and generated schema for Wallet initialization requests. */
+export const initializeWalletRequestSchema = namedSchema<InitializeWalletRequest>(
+  'InitializeWalletRequest',
+  objectNode({ passphrase: stringNode() }, { optional: ['passphrase'] }),
+);
+
+/** Runtime and generated schema for Coco Session start requests. */
+export const startSessionRequestSchema = namedSchema<StartSessionRequest>(
+  'StartSessionRequest',
+  objectNode({ passphrase: stringNode() }, { optional: ['passphrase'] }),
+);
+
+/** Runtime and generated schema for Coco Session stop requests. */
+export const stopSessionRequestSchema = namedSchema<StopSessionRequest>(
+  'StopSessionRequest',
+  objectNode({}),
+);
+
+/** Runtime and generated schema for the sensitive Wallet initialization response. */
+export const initializeWalletResponseSchema = namedSchema<InitializeWalletResponseDocument>(
+  'InitializeWalletResponse',
+  objectNode({
+    generatedMnemonic: stringNode(),
+    status: lifecycleStatusNode,
+  }),
+);
+
+/** Maps runtime-owned lifecycle state to its safe v1 representation. */
+export function toLifecycleStatusDocument(
+  status: CocodStatus,
+  daemonVersion: string,
+): LifecycleStatusDocument {
+  return {
+    daemon: { version: daemonVersion, interfaceVersion: INTERFACE_VERSION },
+    wallet: status.wallet ? { configuredAt: status.wallet.configuredAt } : null,
+    seedAccess: status.seedAccess,
+    cocoSession: status.cocoSession,
+  };
+}
+
+function namedSchema<T>(name: string, node: SchemaNode): RuntimeSchema<T> {
+  return {
+    name,
+    jsonSchema: node.jsonSchema,
+    parse(value) {
+      return node.parse(value, '$') as T;
+    },
+  };
+}
+
+function literalNode<T extends string | null>(expected: T): SchemaNode {
+  return {
+    jsonSchema: expected === null ? { type: 'null' } : { const: expected },
+    parse(value, path) {
+      if (value !== expected) {
+        throw new Error(`${path} must equal ${JSON.stringify(expected)}`);
+      }
+      return expected;
+    },
+  };
+}
+
+function stringNode(options: { format?: 'date-time'; pattern?: string } = {}): SchemaNode {
+  return {
+    jsonSchema: { type: 'string', ...options },
+    parse(value, path) {
+      if (typeof value !== 'string') {
+        throw new Error(`${path} must be a string`);
+      }
+      if (options.pattern && !new RegExp(options.pattern).test(value)) {
+        throw new Error(`${path} does not match the required pattern`);
+      }
+      if (
+        options.format === 'date-time' &&
+        (!Number.isFinite(Date.parse(value)) || new Date(value).toISOString() !== value)
+      ) {
+        throw new Error(`${path} must be an RFC 3339 UTC timestamp`);
+      }
+      return value;
+    },
+  };
+}
+
+function booleanNode(): SchemaNode {
+  return {
+    jsonSchema: { type: 'boolean' },
+    parse(value, path) {
+      if (typeof value !== 'boolean') {
+        throw new Error(`${path} must be a boolean`);
+      }
+      return value;
+    },
+  };
+}
+
+function enumNode<const T extends readonly string[]>(values: T): SchemaNode {
+  return {
+    jsonSchema: { enum: values },
+    parse(value, path) {
+      if (typeof value !== 'string' || !values.includes(value)) {
+        throw new Error(`${path} must be one of ${values.join(', ')}`);
+      }
+      return value;
+    },
+  };
+}
+
+function nullableNode(node: SchemaNode): SchemaNode {
+  return unionNode([literalNode(null), node], 'anyOf');
+}
+
+function objectNode(
+  properties: Readonly<Record<string, SchemaNode>>,
+  options: { optional?: readonly string[]; additionalProperties?: boolean } = {},
+): SchemaNode {
+  const optional = new Set(options.optional ?? []);
+  const additionalProperties = options.additionalProperties ?? false;
+  const required = Object.keys(properties).filter((key) => !optional.has(key));
+  return {
+    jsonSchema: {
+      type: 'object',
+      additionalProperties,
+      required,
+      properties: Object.fromEntries(
+        Object.entries(properties).map(([key, node]) => [key, node.jsonSchema]),
+      ),
+    },
+    parse(value, path) {
+      const record = requireRecord(value, path);
+      if (!additionalProperties) {
+        requireOnlyKeys(record, Object.keys(properties), path);
+      }
+      const parsed: Record<string, unknown> = additionalProperties ? { ...record } : {};
+      for (const [key, node] of Object.entries(properties)) {
+        if (!(key in record)) {
+          if (optional.has(key)) {
+            continue;
+          }
+          throw new Error(`${path}.${key} is required`);
+        }
+        parsed[key] = node.parse(record[key], `${path}.${key}`);
+      }
+      return parsed;
+    },
+  };
+}
+
+function unionNode(nodes: readonly SchemaNode[], keyword: 'oneOf' | 'anyOf' = 'oneOf'): SchemaNode {
+  return {
+    jsonSchema: { [keyword]: nodes.map((node) => node.jsonSchema) },
+    parse(value, path) {
+      const matches: unknown[] = [];
+      const errors: unknown[] = [];
+      for (const node of nodes) {
+        try {
+          matches.push(node.parse(value, path));
+        } catch (error) {
+          errors.push(error);
+        }
+      }
+      if (matches.length === 0 || (keyword === 'oneOf' && matches.length !== 1)) {
+        throw new AggregateError(errors, `${path} does not match ${keyword}`);
+      }
+      return matches[0];
+    },
+  };
+}
+
+function requireRecord(value: unknown, label: string): Record<string, unknown> {
+  if (typeof value !== 'object' || value === null || Array.isArray(value)) {
+    throw new Error(`${label} must be an object`);
+  }
+  return value as Record<string, unknown>;
+}
+
+function requireOnlyKeys(
+  value: Record<string, unknown>,
+  allowedKeys: readonly string[],
+  label: string,
+): void {
+  const allowed = new Set(allowedKeys);
+  if (Object.keys(value).some((key) => !allowed.has(key))) {
+    throw new Error(`${label} contains an unexpected field`);
+  }
+}
